requirements.txt
asgiref==3.2.3
astroid==2.3.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-cors-headers==3.2.1
django-countries==5.5
django-import-export==2.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.8.4
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
PyYAML==5.3
six==1.14.0
sqlparse==0.3.0
tablib==1.1.0
typed-ast==1.4.1
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
